Fixed issue with double in MS Azure Stoarage.

+++ src/Lykke.Pkg.AzureRepositories/AssertPairHistoryRepository.cs
@@ -21,8 +21,8 @@ namespace Lykke.AzureRepositories
             get => RowKey;
             set => RowKey = value;
         }
-        public float Bid { get; set; }
-        public float Ask { get; set; }
+        public double Bid { get; set; }
+        public double Ask { get; set; }
         public int Accuracy { get; set; }
         public DateTime StoredTime { get; set; }
